add script for login in js, welcome template

+++ invoices/models.py
@@ -115,7 +115,7 @@ class InvoiceItems(models.Model):
     period_from = models.DateField(verbose_name='Okres od', null=False)
     period_to = models.DateField(verbose_name='Okres do', null=False)
     measurementSystemNumber = models.CharField(verbose_name='Nr. licznika', null=True, blank=True, max_length=15)
-    counterReading = models.IntegerField(verbose_name='Stan licznika', null=True, blank=True)
+    counterReading = models.DecimalField(verbose_name='Stan licznika', max_digits=10, decimal_places=2, null=True, blank=True)
     consumption = models.IntegerField(verbose_name='Zużycie', null=True, blank=True)
     unit = models.ForeignKey(to=Unit, on_delete=models.CASCADE, verbose_name='Jednostka', related_name=related_name)
     section = models.ForeignKey(to=Section, null=True, blank=True, on_delete=models.CASCADE, verbose_name='Rozdział',
